[Refactor] : BeforeSecondsStageCard 에서 쓰이던 내부 컴포넌트를 각각 분리
1. 하위 컴포넌트들은 추후에 hook을 사용해야 하기 때문에 클라이언트 컴포넌트로서 활용하기 위함임
2. 이거 하나 하자고 BeforeSecondStageCard 컴포넌트 자체를 SC에서 CC로 전환하기에는 아쉬움
